split vaultStrategyModal by product
