+++ src/components/Modal/VaultStrategyModal/compounder/index.ts
@@ -1,1 +1,1 @@
-export * from "./CompounderRightPanel"
+export * from "./CompounderVaultStrategyModal"
