+++ src/components/Modal/VaultStrategyModal/concentrator/index.ts
@@ -1,1 +1,1 @@
-export * from "./ConcentratorRightPanel"
+export * from "./ConcentratorVaultStrategyModal"
