+++ src/components/Modal/VaultStrategyModal/index.ts
@@ -2,6 +2,3 @@ export * from "./compounder"
 export * from "./concentrator"
 export * from "./lib"
 export * from "./managedVaults"
-export * from "./VaultStrategyLeftPanel"
-export * from "./VaultStrategyModal"
-export * from "./VaultStrategyRightPanel"
