+++ src/components/Modal/VaultStrategyModal/managedVaults/index.ts
@@ -1,3 +1,2 @@
 export * from "./ManagedVaultsActivityTable"
-export * from "./ManagedVaultsLeftPanel"
-export * from "./ManagedVaultsRightPanel"
+export * from "./ManagedVaultsStrategyModal"
